+++ docs/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ### v0.1.1
+- Set up command system in prompt
+- Added commands: `About`, `Help`
 - Added expressions: `Boolean`, `Variable`
 - Added functions:
   - Arithmetic: `Decimal`, `Exponent`, `Modulo`, `Positive`, `Negative`
@@ -8,6 +10,8 @@
   - Counting: `Factorial`
 - Modified existing functions to support new expression types
 - Improved engine independence of expression and function tests
+- Added tests: `Commands`
+- Removed tests: `Banner`
 ### v0.1.0
 - Set up engine framework, bundling system using Rollup and testing system using Jest
 - Set up prompt and web examples